Refactoring: Presentation logic removed from Use Cases
