+++ app/src/main/java/com/github/kulebin/myfoursquareapp/useCase/OnInteractionCallback.java
@@ -1,0 +1,9 @@
+package com.github.kulebin.myfoursquareapp.useCase;
+
+interface OnInteractionCallback {
+
+    void onStart();
+
+    void onError(Exception e);
+
+}
