+++ app/src/main/java/com/github/kulebin/myfoursquareapp/view/VenueDetailContract.java
@@ -2,14 +2,12 @@ package com.github.kulebin.myfoursquareapp.view;
 
 public interface VenueDetailContract {
 
-    interface View extends BaseContract.View{
+    interface View extends BaseContract.View {
 
         void displayData(CompleteVenueDisplayData venueToShowData);
     }
 
-    interface Presentation extends BaseContract.Presentation {
-
-        void presentVenueToShowData(CompleteVenueDisplayData venueToShowData);
+    interface Presentation {
 
         void onViewCreated(String venueId);
