+++ app/src/main/java/com/github/kulebin/myfoursquareapp/view/VenueListPresenter.java
@@ -1,14 +1,11 @@
-package com.github.kulebin.myfoursquareapp.presenter;
+package com.github.kulebin.myfoursquareapp.view;
 
 import com.github.kulebin.myfoursquareapp.adapter.VenueListAdapter;
 import com.github.kulebin.myfoursquareapp.useCase.ShowVenueListUseCase;
-import com.github.kulebin.myfoursquareapp.view.VenueDisplayData;
-import com.github.kulebin.myfoursquareapp.view.VenueItemView;
-import com.github.kulebin.myfoursquareapp.view.VenueListContract;
 
 import java.util.List;
 
-public class VenueListPresenter implements VenueListContract.Presentation {
+public class VenueListPresenter implements VenueListContract.Presentation, ShowVenueListUseCase.IRecipient {
 
     private List<VenueDisplayData> mVenueDisplayList;
     private final VenueListContract.View mView;
@@ -22,6 +19,7 @@ public class VenueListPresenter implements VenueListContract.Presentation {
     @Override
     public void presentVenueToShowData(final List<VenueDisplayData> venueToShowData) {
         this.mVenueDisplayList = venueToShowData;
+        mView.showProgress(false);
         mVenueListAdapter.notifyDataSetChanged();
     }
 
@@ -47,13 +45,14 @@ public class VenueListPresenter implements VenueListContract.Presentation {
     }
 
     @Override
-    public void onError(final Exception e) {
-        mView.showError(e.getMessage());
+    public void onStart() {
+        mView.showProgress(true);
     }
 
     @Override
-    public void setProgress(final boolean isVisible) {
-        mView.showProgress(isVisible);
+    public void onError(final Exception e) {
+        mView.showProgress(false);
+        mView.showError(e.getMessage());
     }
 
     @Override
