+++ app/src/main/res/layout/fragment_venue_detail.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:padding="12dp"
-    tools:context="com.github.kulebin.myfoursquareapp.VenueListFragment">
+    tools:context="com.github.kulebin.myfoursquareapp.view.VenueListFragment">
 
     <LinearLayout
         android:layout_width="match_parent"
