+++ app/src/main/res/layout/fragment_venue_list.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.github.kulebin.myfoursquareapp.VenueListFragment">
+    tools:context="com.github.kulebin.myfoursquareapp.view.VenueListFragment">
 
     <android.support.v7.widget.RecyclerView
         android:id="@+id/recycler_view_venue_list"
@@ -15,7 +15,7 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         app:layout_behavior="@string/appbar_scrolling_view_behavior"
-        tools:context=".VenueListFragment">
+        tools:context=".view.VenueListFragment">
 
     </android.support.v7.widget.RecyclerView>
